Corrected a problem with the OpenGUI post-build event not working correctly when dealing with paths that contain spaces.
Removed OpenGUI.suo, again. It really is an unnecessary file.

Added OgreFusion, which is a complete set of classes to support using OpenGUI in Ogre projects. This covers the rendering of OpenGUI to Ogre, as well as using Ogre's built in resource system. Currently only the resource class is finished, the rest is just being committed for completeness.